+++ examples/timeout.rs
@@ -1,9 +1,15 @@
 //! Timeout example - this demonstrates running one task with a timeout continuously.
 
+use core::time::Duration;
 use std::task::Poll;
 
-use bevy::{app::PanicHandlerPlugin, log::LogPlugin, prelude::*};
-use bevy_async_task::{Duration, TimedAsyncTask, TimedTaskRunner, TimeoutError, sleep};
+use bevy::app::PanicHandlerPlugin;
+use bevy::log::LogPlugin;
+use bevy::prelude::*;
+use bevy_async_task::TimedAsyncTask;
+use bevy_async_task::TimedTaskRunner;
+use bevy_async_task::TimeoutError;
+use bevy_async_task::sleep;
 
 fn system_does_timeout(mut task_executor: TimedTaskRunner<'_, ()>) {
     if task_executor.is_idle() {
@@ -14,7 +20,7 @@ fn system_does_timeout(mut task_executor: TimedTaskRunner<'_, ()>) {
 
     match task_executor.poll() {
         Poll::Ready(Err(TimeoutError)) => {
-            info!("Timeout on A!");
+            info!("Timeout on A! (expected)");
         }
         Poll::Pending => {
             // Waiting...
@@ -33,6 +39,7 @@ fn system_doesnt_timeout(mut task_executor: TimedTaskRunner<'_, u32>, mut counte
                 next
             }
         };
+        let task = TimedAsyncTask::new(std::time::Duration::from_millis(2_147_483_647_u64), task);
         task_executor.start(task);
         info!("Started B!");
     }
@@ -44,7 +51,7 @@ fn system_doesnt_timeout(mut task_executor: TimedTaskRunner<'_, u32>, mut counte
         Poll::Pending => {
             // Waiting...
         }
-        e => panic!("{e:?}"),
+        e => panic!("This shouldn't timeout!: {e:?}"),
     }
 }