+++ src/task_runner.rs
@@ -1,18 +1,23 @@
-use std::{
-    ops::{Deref, DerefMut},
-    sync::atomic::Ordering,
-    task::Poll,
-};
+use std::ops::Deref;
+use std::ops::DerefMut;
+use std::sync::atomic::Ordering;
+use std::task::Poll;
 
-use bevy_ecs::{
-    component::Tick,
-    system::{ExclusiveSystemParam, ReadOnlySystemParam, SystemMeta, SystemParam},
-    world::{World, unsafe_world_cell::UnsafeWorldCell},
-};
+use bevy_ecs::component::Tick;
+use bevy_ecs::system::ExclusiveSystemParam;
+use bevy_ecs::system::ReadOnlySystemParam;
+use bevy_ecs::system::SystemMeta;
+use bevy_ecs::system::SystemParam;
+use bevy_ecs::world::World;
+use bevy_ecs::world::unsafe_world_cell::UnsafeWorldCell;
 use bevy_platform::cell::SyncCell;
-use bevy_tasks::{AsyncComputeTaskPool, ConditionalSend};
+use bevy_tasks::AsyncComputeTaskPool;
+use bevy_tasks::ConditionalSend;
 
-use crate::{AsyncReceiver, AsyncTask, TimedAsyncTask, TimeoutError};
+use crate::AsyncReceiver;
+use crate::AsyncTask;
+use crate::TimedAsyncTask;
+use crate::TimeoutError;
 
 /// A Bevy [`SystemParam`] to execute async tasks in the background.
 #[derive(Debug)]